UMCS-172(refactor) move backwards compatibility test for hire purchase into a separate test class.
