+++ test/integration/PaymentTypes/HirePurchaseTest.php
@@ -1,0 +1,143 @@
+<?php
+
+/** @noinspection PhpUnhandledExceptionInspection */
+/** @noinspection PhpDocMissingThrowsInspection */
+/**
+ * This class defines integration tests to verify backwards compatibility with deprecated hire-purchase-direct-debit resource.
+ *
+ * Copyright (C) 2020 - today Unzer E-Com GmbH
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * @link  https://docs.unzer.com/
+ *
+ * @package  UnzerSDK\test\integration\PaymentTypes
+ */
+namespace UnzerSDK\test\integration\PaymentTypes;
+
+use UnzerSDK\Exceptions\UnzerApiException;
+use UnzerSDK\Resources\AbstractUnzerResource;
+use UnzerSDK\Resources\PaymentTypes\InstallmentSecured;
+use UnzerSDK\Resources\TransactionTypes\Charge;
+use UnzerSDK\test\BaseIntegrationTest;
+use function count;
+
+class HirePurchaseTest extends BaseIntegrationTest
+{
+    /**
+     * Verify, backwards compatibility regarding fetching payment type and map it to invoice secured class.
+     *
+     * @test
+     */
+    public function hddTypeShouldBeFetchable(): InstallmentSecured
+    {
+        // Mock a hdd Type
+        $date = $this->getTodaysDateString();
+        $requestData =
+            [
+                "iban" =>"DE89370400440532013000",
+                "bic" => "COBADEFFXXX",
+                "accountHolder" => "Max Mustermann",
+                "invoiceDueDate" => $date,
+                "numberOfRates" => 3,
+                "invoiceDate" => $date,
+                "dayOfPurchase" => $date,
+                "orderDate" => $date,
+                "totalPurchaseAmount" => 119,
+                "totalInterestAmount" => 0.96,
+                "totalAmount" => 119.96,
+                "effectiveInterestRate" => 4.99,
+                "nominalInterestRate" => 4.92,
+                "feeFirstRate" => 0,
+                "feePerRate" => 0,
+                "monthlyRate" => 39.99,
+                "lastRate" => 39.98,
+        ];
+
+        $payload = json_encode($requestData);
+        $hddMock = $this->getMockBuilder(InstallmentSecured::class)
+            ->setMethods(['getUri', 'jsonSerialize'])
+            ->getMock();
+        $hddMock->method('getUri')->willReturn('/types/hire-purchase-direct-debit');
+        $hddMock->method('jsonSerialize')->willReturn($payload);
+
+        // When
+        /** @var InstallmentSecured $hddMock */
+        $this->unzer->createPaymentType($hddMock);
+        $this->assertRegExp('/^s-hdd-[.]*/', $hddMock->getId());
+
+        // Then
+        $fetchedType = $this->unzer->fetchPaymentType($hddMock->getId());
+        $this->assertInstanceOf(InstallmentSecured::class, $fetchedType);
+        $this->assertRegExp('/^s-hdd-[.]*/', $fetchedType->getId());
+
+        return $fetchedType;
+    }
+
+    /**
+     * Verify fetched hdd type can be authorized and charged
+     *
+     * @test
+     * @depends hddTypeShouldBeFetchable
+     *
+     * @param InstallmentSecured $hddType fetched hdd type.
+     *
+     * @return AbstractUnzerResource|Charge
+     *
+     * @throws UnzerApiException
+     */
+    public function hddTypeAuthorizeAndCharge(InstallmentSecured $hddType)
+    {
+        $customer = $this->getMaximumCustomer();
+        $basket = $this->createBasket();
+
+        $auth = $hddType->authorize(119.00, 'EUR', 'https://unzer.com', $customer, null, null, $basket);
+        $charge = $auth->getPayment()->charge();
+        $this->assertNotNull($auth);
+        $this->assertNotEmpty($auth->getId());
+        $this->assertTrue($auth->isSuccess());
+
+        return $charge;
+    }
+
+    /**
+     * Verify fetched hdd payment can be shipped.
+     *
+     * @test
+     * @depends hddTypeAuthorizeAndCharge
+     */
+    public function hddTypeShouldBeShippable(Charge $hddCharge)
+    {
+        $invoiceId = 'i' . self::generateRandomId();
+        $ship = $this->unzer->ship($hddCharge->getPayment(), $invoiceId);
+
+        $this->assertNotNull($ship);
+        return $hddCharge;
+    }
+
+    /**
+     * Verify full cancel of charged HP after shipment.
+     *
+     * @test
+     * @depends hddTypeAuthorizeAndCharge
+     */
+    public function hddChargeCanBePartiallyCancledBeforeShipment(Charge $hddCharge): void
+    {
+        $payment = $hddCharge->getPayment();
+
+        $cancel1 = $payment->cancelAmount(66, null, null, 60, 6);
+        $cancel2 = $payment->cancelAmount(43, null, null, 40, 3);
+        $this->assertGreaterThan(0, count($cancel1));
+        $this->assertGreaterThan(0, count($cancel2));
+    }
+}
